requirements.txt
flask==2.0.1
tensorflow==2.15.0
numpy==1.24.3
Pillow==9.5.0 